+++ hack/bundle-gen.py
@@ -12,15 +12,8 @@ import sys
 import tempfile
 import urllib3
 import yaml
-import re
 
-# HIVE_VERSION_PREFIX is the prefix of the Hive version that will be constructed by
-# this script if we get (or default to) a --branch that tracks master. Otherwise we
-# will use the branch to calculate the prefix (e.g. `--branch ocm-2.3` => '2.3').
-# The version used for images and bundles will be:
-# "{prefix}.{number of commits}-{git hash}"
-# e.g. "v1.2.3187-18827f6"
-HIVE_VERSION_PREFIX = "1.2"
+import version
 
 HIVE_REPO_DEFAULT = "git@github.com:openshift/hive.git"
 
@@ -35,21 +28,11 @@ REGISTRY_AUTH_FILE_DEFAULT = "{}/.docker/config.json".format(os.environ["HOME"])
 
 COMMUNITY_OPERATORS_HIVE_PKG_URL = "https://raw.githubusercontent.com/redhat-openshift-ecosystem/community-operators-prod/main/operators/hive-operator/hive.package.yaml"
 
-CHANNEL_DEFAULT = "alpha"
-HIVE_BRANCH_DEFAULT = "master"
-
 SUBPROCESS_REDIRECT = subprocess.DEVNULL
 
-# Match things like 'ocm-2.5-mce-2.0' or origin/ocm-2.5-mce-2.0, capturing:
-# 1: 'origin/'
-# 2: '2.5' (used for the bundle semver prefix)
-# 3: 'mce-2.0' (used for the channel name)
-MCE_BRANCH_RE = re.compile("^([^/]+/)?ocm-(\d+\.\d+)-(mce-\d+\.\d+)$")
-# Match things like 'ocm-2.3' or 'origin/ocm-2.3', capturing:
-# 1: 'origin/'
-# 2: 'ocm-2.3'
-# 3: '2.3'
-OCM_BRANCH_RE = re.compile("^([^/]+/)?(ocm-(\d+\.\d+))$")
+HIVE_BRANCH_DEFAULT = version.HIVE_BRANCH_DEFAULT
+
+CHANNEL_DEFAULT = version.CHANNEL_DEFAULT
 
 
 def get_params():
@@ -82,7 +65,7 @@ def get_params():
                                 channel named `mce-M.N`.""".format(
             HIVE_BRANCH_DEFAULT,
             HIVE_BRANCH_DEFAULT,
-            HIVE_VERSION_PREFIX,
+            version.HIVE_VERSION_PREFIX,
             CHANNEL_DEFAULT,
             HIVE_BRANCH_DEFAULT,
         ),
@@ -181,25 +164,6 @@ def build_and_push_image(registry_auth_file, hive_version, dry_run, build_engine
     subprocess.run(
         build['push'].format(registry_auth_arg, container_name).split(),
         check=True,
-    )
-
-
-# gen_hive_version generates and returns the hive version eg. "1.2.3187-18827f6"
-def gen_hive_version(repo, commit_hash, prefix):
-    # sha is the first 7 characters of commit_hash
-    sha = commit_hash[0:7]
-
-    # num commits is the number of git commits counted from the first commit to the provided commit_hash
-    # this is the equivalent of running:
-    # git rev-list `git rev-list --parents HEAD | egrep "^[a-f0-9]{40}$"`..HEAD --count
-    parent = repo.git.rev_list("--max-parents=0", commit_hash)
-    num_commits = repo.git.rev_list(
-        "--count",
-        "{parent}..{commit_hash}".format(parent=parent, commit_hash=commit_hash),
-    )
-
-    return "{prefix}.{commits}-{sha}".format(
-        prefix=prefix, commits=num_commits, sha=sha
     )
 
 
@@ -494,116 +458,6 @@ def open_pr(
     print()
 
 
-def parse_branch_name(branch_name):
-    """Split up a branch name of the form 'ocm-X.Y[-mce-M.N].
-
-    :param branch_name: A branch name. If of the form [remote/]ocm-X.Y[-mce-M.N] we will parse
-        it as noted below; otherwise the first return will be False.
-    :return parsed (bool): True if the branch_name was parseable; False otherwise.
-    :return remote (str): If parsed and the branch_name contained a remote/ prefix, it is
-        returned here; otherwise this is the empty string.
-    :return prefix (str): Two-digit semver prefix of the bundle to be generated. If the branch
-        name is of the form [remote/]ocm-X.Y, this will be X.Y; if of the form
-        [remote/]ocm-X.Y-mce-M.N it will be M.N. If not parseable, it will be the empty string.
-    :return channel (str): The name of the channel in which we'll include the bundle. If the
-        branch name is of the form [remote/]ocm-X.Y, this will be ocm-X.Y; if of the form
-        [remote/]ocm-X.Y-mce-M.N it will be mce-M.N. If not parseable, it will be the empty
-        string.
-    """
-    m = MCE_BRANCH_RE.match(branch_name)
-    if m:
-        return True, m.group(1), m.group(2), m.group(3)
-    m = OCM_BRANCH_RE.match(branch_name)
-    if m:
-        return True, m.group(1), m.group(3), m.group(2)
-    return False, '', '', ''
-
-
-def process_branch(hive_repo, branch_arg):
-    """Validate and process the input (or default) branch.
-
-    :param hive_repo: The git.Repo for the local hive clone.
-    :param branch_arg: The string commit-ish input (or defaulted) via the --branch arg.
-    :return commit_hash: The canonical full-length sha of the commit corresponding to branch_arg.
-    :return prefix: The two-digit semver prefix of the bundle version to use. If branch_arg is
-        (a descendant of) master, we'll use HIVE_VERSION_PREFIX. If not, and it names an 'ocm-X.Y'
-        branch, we'll use X.Y.
-    :return channels: List of string channel names to update. If branch_arg is (a descendant of)
-        master, this will include 'alpha'. If branch_arg names an 'ocm-X.Y' branch, it will include
-        'ocm-X.Y'.
-    :raise: If we get a branch_arg that's invalid (doesn't correspond to a real commit in hive_repo),
-        or that's not 'ocm-X.Y' or (a descendant of) master.
-    """
-    # We're cloning the hive repo, so there's no local branch for ocm-X.Y[-mce-M.N]; but we don't want to
-    # force the user to say 'origin/ocm-X.Y[-mce-M.N]'. Accommodate...
-    parsed, remote, _, _ = parse_branch_name(branch_arg)
-    if parsed and not remote:
-        branch_arg = "origin/{}".format(branch_arg)
-
-    # This will raise an exception if there's no such commit-ish
-    commit_hash = hive_repo.rev_parse(branch_arg).hexsha
-
-    # We need to know where master is, even if we're processing a different branch. However, if
-    # we've cloned from something other than HIVE_REPO_DEFAULT, there may not be a local `master`
-    # branch. So we may need to try origin and upstream to find it.
-    for remote in ("", "origin/", "upstream/"):
-        master_branch = remote + HIVE_BRANCH_DEFAULT
-        print("Trying to find master at {}".format(master_branch))
-        try:
-            master_hash = hive_repo.rev_parse(master_branch).hexsha
-        except git.BadName:
-            print("Couldn't find master at `{}`".format(master_branch))
-        else:
-            print("Found master at `{}`".format(master_branch))
-            break
-    else:
-        print("Couldn't find master branch!")
-        sys.exit(1)
-
-    is_master_ancestor = hive_repo.git.rev_list(
-        "--ancestry-path", "{}..{}".format(commit_hash, master_hash)
-    )
-
-    # Was (a commit on) an ocm-X.Y[-mce-M.N] branch specified?
-    # This will find all such branch names and add them to the channel list.
-    # However, we'll only support >1 entry in this list if we're tracking
-    # master, because otherwise which would we use to compute the semver?
-    channels = []
-    maybe_prefix = None
-    for ref in hive_repo.refs:
-        parsed, remote, prefix, channel = parse_branch_name(ref.name)
-        # Only pay attention to remote refs
-        if parsed and remote == "origin/":
-            if hive_repo.rev_parse(ref.name).hexsha == commit_hash:
-                channels.append(channel)
-                # We'll only use this if we end up with one
-                # entry in this list AND we're not tracking master
-                maybe_prefix = prefix
-
-    if commit_hash == master_hash or (is_master_ancestor and not maybe_prefix):
-        prefix = HIVE_VERSION_PREFIX
-        if commit_hash == master_hash:
-            channels.append(CHANNEL_DEFAULT)
-    else:
-        if len(channels) > 1:
-            raise ValueError(
-                "Found more than one channel candidate at {}: {}.\n".format(
-                    branch_arg, channels
-                )
-                + "I can't handle this unless they're direct descendants of master -- "
-                + "which one would I use as the semver base?"
-            )
-        prefix = maybe_prefix
-
-    if not channels or not prefix:
-        raise ValueError(
-            "Can't make sense of branch {}: expected {}, a direct descendant thereof, 'ocm-X.Y', or 'ocm-X.Y-mce-M.N'.".format(
-                branch_arg, HIVE_BRANCH_DEFAULT
-            )
-        )
-    return commit_hash, prefix, channels
-
-
 if __name__ == "__main__":
     args = get_params()
 
@@ -618,7 +472,7 @@ if __name__ == "__main__":
 
     hive_repo = git.Repo(hive_repo_dir.name)
 
-    hive_commit, hive_version_prefix, update_channels = process_branch(
+    hive_commit, hive_version_prefix, update_channels = version.process_branch(
         hive_repo, args.branch
     )
 
@@ -640,7 +494,7 @@ if __name__ == "__main__":
         print("Failed to checkout {}".format(hive_commit))
         raise
 
-    hive_version = gen_hive_version(hive_repo, hive_commit, hive_version_prefix)
+    hive_version = version.gen_hive_version(hive_repo, hive_commit, hive_version_prefix)
     prev_version = get_previous_version(channel)
     if hive_version == prev_version:
         raise ValueError("Version {} already exists upstream".format(hive_version))
